% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/agglomerative_clustering.R
\name{agglomerative_clustering}
\alias{agglomerative_clustering}
\title{Agglomerative Hierarchical Clustering}
\usage{
agglomerative_clustering(data, proximity = "MIN", method = "euclidean", p = 3)
}
\arguments{
\item{data}{a set of observations, presented as a matrix-like object where
every row is a new observation.}

\item{proximity}{the proximity definition to be used. This should be one
of \code{"MIN"} (single linkage), \code{"MAX"} (complete linkage),
\code{"AVG"} (average linkage).}

\item{method}{the distance definition to be used. Check \code{\link[=distance]{distance()}} for the
available methods.}

\item{p}{the exponent of the Minkowski distance.}
}
\value{
An \code{\link[=hclust]{hclust()}} object which describes the tree produced by the
clustering process.
}
\description{
Hierarchical agglomerative cluster analysis on a set of
observations
}
\details{
This function performs a hierarchical cluster analysis for the
\eqn{n} objects being clustered. The definition of a set of clusters using
this method follows a \eqn{n} step process, which repeats until a single
cluster remains:

\enumerate{
\item Initially, each object is assigned to its own cluster. The matrix
of distances between clusters is computed. This is done according to the
specified \code{method}.
\item The two clusters with closest proximity be joined together and the
proximity matrix will be updated. This is done according to the specified
\code{proximity}. This step is repeated until a single cluster remains.
}

The definitions of \code{proximity} considered by this function are:

\describe{
\item{\code{MIN}}{\eqn{\min\left\{d(x,y):x\in A,y\in B\right\}}. Defines the
proximity between two clusters as the distance between the closest objects
among the two clusters. It produces clusters where each object is closest to
at least one other object in the same cluster. It is also known as
\strong{SLINK} or \strong{single-link}.}
\item{\code{MAX}}{\eqn{\max\left\{d(x,y):x\in A,y\in B\right\}}. Defines the
proximity between two clusters as the distance between the furthest objects
among the two clusters. It is also known as \strong{CLINK} or
\strong{complete-link}.}
\item{\code{AVG}}{\eqn{\frac{1}{\left|A\right|\cdot\left|B\right|}
 \sum_{x\in A}\sum_{y\in B} d(x,y)}. Defines the proximity between two
clusters as the average distance between every pair of objects, one from
each cluster. It is also known as \strong{UPGMA} or \strong{average-link}.}
}
}
\examples{
### Example 1: Violent crime rates by US state

ac <- agglomerative_clustering(USArrests, proximity = "AVG")
plot(ac, hang = -1)

### Example 2: Iris flower dataset

ac <- agglomerative_clustering(iris[, 1:4], proximity = "AVG")

# Dendrogram
plot(ac, hang = -1, label = iris$Species)
rect.hclust(ac, k = 3, border = 1:3)

# Scatterplot
plot(iris, col = cutree(ac, k = 3))

}
\author{
Eduardo Ruiz Sabajanes, \email{eduardo.ruizs@edu.uah.es}
}
