% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{db1}
\alias{db1}
\title{Test Database 1}
\format{
\subsection{\code{db1}}{

A data frame with 500 rows and 2 columns.

The data points form two concentric circles.
}
}
\usage{
db1
}
\description{
Test Database 1
}
\keyword{datasets}
