% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{db2}
\alias{db2}
\title{Test Database 2}
\format{
\subsection{\code{db2}}{

A data frame with 500 rows and 2 columns.

The data points form two moons.
}
}
\usage{
db2
}
\description{
Test Database 2
}
\keyword{datasets}
