% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{db3}
\alias{db3}
\title{Test Database 3}
\format{
\subsection{\code{db3}}{

A data frame with 500 rows and 2 columns.

The data points form three overlapping elliptical clusters of varying
densities.
}
}
\usage{
db3
}
\description{
Test Database 3
}
\keyword{datasets}
