% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{db4}
\alias{db4}
\title{Test Database 4}
\format{
\subsection{\code{db4}}{

A data frame with 500 rows and 2 columns.

The data points form three diagonal parallel segments.
}
}
\usage{
db4
}
\description{
Test Database 4
}
\keyword{datasets}
