% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{db5}
\alias{db5}
\title{Test Database 5}
\format{
\subsection{\code{db5}}{

A data frame with 500 rows and 2 columns.

The data points form three non-overlapping circular clusters of similar
density.
}
}
\usage{
db5
}
\description{
Test Database 5
}
\keyword{datasets}
