% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{db6}
\alias{db6}
\title{Test Database 6}
\format{
\subsection{\code{db6}}{

A data frame with 500 rows and 2 columns.

The data points are uniformly distributed on the plane.
}
}
\usage{
db6
}
\description{
Test Database 6
}
\keyword{datasets}
