% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dbscan.R
\name{dbscan}
\alias{dbscan}
\title{Divisive Hierarchical Clustering}
\usage{
dbscan(data, eps = 0.5, min_pts = 5, method = "euclidean", p = 2)
}
\arguments{
\item{data}{a set of observations, presented as a matrix-like object where
every row is a new observation.}

\item{eps}{how close two observations have to be to be considered neighbors.}

\item{min_pts}{the minimum amount of neighbors for a region to be considered
dense.}

\item{method}{the distance definition to be used. Check \code{\link[=distance]{distance()}} for the
available methods.}

\item{p}{the power of the Minkowski distance to be used. Only used if
\code{method = "minkowski"}.}
}
\value{
A \code{\link[=kmeans]{kmeans()}} object.
}
\description{
Hierarchical divisive cluster analysis on a set of observations
}
\examples{
# a 2-dimensional example
x <- rbind(matrix(rnorm(100, sd = 0.3), ncol = 2),
           matrix(rnorm(100, mean = 1, sd = 0.3), ncol = 2))
colnames(x) <- c("x", "y")
(cl <- dbscan(x, 0.5, 5))
plot(x, col = cl, pch = 20)

}
\author{
Eduardo Ruiz Sabajanes, \email{eduardo.ruizs@edu.uah.es}
}
