% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/divisive_clustering.R
\name{divisive_clustering}
\alias{divisive_clustering}
\title{Divisive Hierarchical Clustering}
\usage{
divisive_clustering(data, max_iterations = 10)
}
\arguments{
\item{data}{a set of observations, presented as a matrix-like object where
every row is a new observation.}

\item{max_iterations}{the maximum number of iterations to use in the calls to
\code{\link[=kmeans]{kmeans()}} when the algorithm is looking for a bipartition of a cluster.}
}
\value{
An \code{\link[=hclust]{hclust()}} object which describes the tree produced by the
clustering process.
}
\description{
Hierarchical divisive cluster analysis on a set of observations
}
\author{
Eduardo Ruiz Sabajanes, \email{eduardo.ruizs@edu.uah.es}
}
