% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/kmeans.R
\name{kmeans}
\alias{kmeans}
\title{K-Means Clustering}
\usage{
kmeans(data, centers, max_iterations = 10, method = "euclidean", p = 3)
}
\arguments{
\item{data}{a set of observations, presented as a matrix-like object where
every row is a new observation.}

\item{centers}{either the number of clusters or a set of initial cluster
centers. If a number, a random set of rows in x is chosen as the initial
centers.}

\item{max_iterations}{the maximum number of iterations allowed.}

\item{method}{the distance definition to be used. Check \code{\link[=distance]{distance()}} for the
available methods.}

\item{p}{the exponent of the Minkowski distance.}
}
\value{
A \code{\link[=kmeans]{kmeans()}} object.
}
\description{
Perform K-Means clustering on a data matrix.
}
\author{
Eduardo Ruiz Sabajanes, \email{eduardo.ruizs@edu.uah.es}
}
